requirements.txt
Pillow==10.0.1
tqdm==4.66.1